# Speed up square matrix times vector over GF(2)