@grhkm21 2024-02-21
Just a note, your timing doesn't show anything, since the dimensions are different (one's with $n = 100$, one is with $n = 10$)