@yyyyx4 2024-02-21
Oh! Good catch. But there is still a significant speedup: The reduction in runtime is about $50\\,\\%$ for $10\times10$ matrices and about $20\\,\\%$ for $100\times100$ matrices.