@kedlaya 2024-02-21
I fixed the examples to both use 10 by 10 matrices, as indeed this problem is more acute with small matrices. 

There seems to be a fixed amount of overhead with matrix/vector creation which is rampant throughout Sage; a broader fix for that would obviate the need for such trickery.